pyproject.toml
[project]
  dependencies = [
    "pytest",
    "pyright",
    "ruff",
    "notify-py>=0.3.43",
    "loguru>=0.6.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "tabulate>=0.9.0",
  ]
  description = "Root .claude directory."
  name = "dot-claude"
  requires-python = ">=3.12"
  version = "1.0.0"

[project.scripts]
  cc-notifier = "hooks.cc_notifier.cli:cli"

[tool.setuptools]
  packages = ["hooks.cc_notifier"]

[dependency-groups]
  dev = [
    "pytest-mock>=3.15.1",
    "pytest-cov>=6.0.0",
  ]
